# api-spec.yaml
openapi: 3.1.0
info:
  title: Authentik Role UI Backend
  version: 1.0.0
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    GroupLink:
      type: object
      required: [name]
      properties:
        name:
          type: string
...
    GroupRoleSplit:
      description: Members of a group split by their role.
      type: object
      required: [leader, manager, member]
      properties:
        leader:
          type: array
          items:
            $ref: '#/components/schemas/UserLink'
        manager:
          type: array
          items:
            $ref: '#/components/schemas/UserLink'
        member:
          type: array
          items:
            $ref: '#/components/schemas/UserLink'
...
    GoogleSyncConfig:
      type: object
      properties:
        recursive_name:
          type: string
          pattern: '^[a-z0-9.]+$'
        direct_name:
          type: string
          pattern: '^[a-z0-9.]+$'
    Group:
      type: object
      required: [name, members, children, parents]
      properties:
        name:
          type: string
        members:
          $ref: '#/components/schemas/GroupRoleSplit'
        children:
          type: array
          items:
            $ref: '#/components/schemas/GroupLink'
        parents:
          type: array
          items:
            $ref: '#/components/schemas/GroupLink'
        color:
          type: string
        google_sync:
          $ref: '#/components/schemas/GoogleSyncConfig'
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Caller lacks the required role in this group.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: User or group not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    BadRequest:
      description: Invalid request parameters or body.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  parameters:
    GroupName:
      name: group_name
      in: path
      required: true
      schema:
        type: string
...
paths:
...
  /groups:
    get:
      operationId: getGroups
      summary: List all groups
      responses:
        '200':
          description: All groups.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /groups/{group_name}:
    get:
      operationId: getGroup
      summary: Get a group by name
      parameters:
        - $ref: '#/components/parameters/GroupName'
      responses:
        '200':
          description: Group detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: disbandGroup
      summary: Disband a group
      description: >
        Permanently deletes the group. Requires caller to be the group's **leader**.
        Fails with 400 if the group has any children — detach or disband them first.
      parameters:
        - $ref: '#/components/parameters/GroupName'
      responses:
        '200':
          description: Group disbanded.
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{group_name}/members:
    post:
      operationId: addMember
      summary: Add a member
      description: >
        Adds a user to the group as a plain member. Idempotent: if the user is
        already a member in any role the call succeeds without changes. Requires
        caller to be **manager or leader**.
      parameters:
        - $ref: '#/components/parameters/GroupName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
      responses:
        '200':
          description: Member added (or was already present).
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /groups/{group_name}/managers:
    post:
      operationId: addManager
      summary: Promote a member to manager
      description: >
        Grants the manager role to an existing group member. Requires caller to be
        the group's **leader**. The user must already be a member; idempotent if
        they are already a manager.
      parameters:
        - $ref: '#/components/parameters/GroupName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
      responses:
        '200':
          description: Manager role granted.
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /groups/{group_name}/leaders:
    post:
      operationId: addLeader
      summary: Promote a member to leader (superuser only)
      description: >
        Grants the leader role to an existing group member. Requires the caller to
        be an authentik superuser operating in superuser mode (`x-as-superuser: true`).
        The user must already be a member of the group. Idempotent if already a leader.
        If the user was a manager, their manager role is removed.
      parameters:
        - $ref: '#/components/parameters/GroupName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
      responses:
        '200':
          description: Leader role granted (or user was already a leader).
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /groups/{group_name}/leader/resign:
    post:
      operationId: resignLeader
      summary: Resign as leader
      description: >
        Transfers leadership to a successor. The successor must be a current
        **manager** of the group; they are promoted to leader and removed from
        the managers list. Requires caller to be the current **leader**.
      parameters:
        - $ref: '#/components/parameters/GroupName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                  description: Username of the manager who will become the new leader.
      responses:
        '200':
          description: Leadership transferred.
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{group_name}/subgroups:
    post:
      operationId: createSubgroup
      summary: Create a new child group
      description: >
        Creates a brand-new group as a child of this one. The calling **leader**
        is automatically added as a member and assigned as leader of the new group.
        Name must be 1–150 characters.
      parameters:
        - $ref: '#/components/parameters/GroupName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 150
      responses:
        '200':
          description: Child group created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupLink'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{group_name}/children:
    post:
      operationId: addChildGroup
      summary: Attach an existing group as a child
      description: >
        Makes an existing group a child of this one. Requires caller to be the
        **leader** of the parent group. Rejected with 400 if the attachment would
        create a cycle in the group hierarchy.
      parameters:
        - $ref: '#/components/parameters/GroupName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [group_name]
              properties:
                group_name:
                  type: string
                  description: Name of the existing group to attach as a child.
      responses:
        '200':
          description: Child attached.
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /groups/{group_name}/name:
    put:
      operationId: renameGroup
      summary: Rename a group
      description: >
        Changes the group's name. Requires caller to be the group's **leader**
        or a superuser in superuser mode. Name must be at least 5 characters.
      parameters:
        - $ref: '#/components/parameters/GroupName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 5
      responses:
        '200':
          description: Group renamed.
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{group_name}/google-sync:
    patch:
      operationId: setGoogleSync
      summary: Set Google Workspace sync config
      description: >
        Sets `recursive_name` and/or `direct_name` for the group's Google Workspace
        sync configuration. Both fields are optional — omit a field to leave it
        unchanged. Requires caller to be the group's **leader**.
      parameters:
        - $ref: '#/components/parameters/GroupName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recursive_name:
                  type: string
                  pattern: '^[a-z0-9.]+$'
                direct_name:
                  type: string
                  pattern: '^[a-z0-9.]+$'
      responses:
        '200':
          description: Sync config updated.
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{group_name}/color:
    put:
      operationId: setGroupColor
      summary: Set group color
      description: >
        Updates the display color of a group. Requires caller to be **manager or
        leader**. Color value is a free-form string (e.g. a hex code like
        `"FF0000"`).
      parameters:
        - $ref: '#/components/parameters/GroupName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [color]
              properties:
                color:
                  type: string
                  example: FF0000
      responses:
        '200':
          description: Color updated.
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Search
  # ---------------------------------------------------------------------------
...